% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{abod_analysis}
\alias{abod_analysis}
\title{abod analysis}
\usage{
abod_analysis(data, abod = NULL, k = 15)
}
\arguments{
\item{data}{the data we want to analyze}

\item{abod}{the abod if it exist}

\item{k}{for the abod}
}
\description{
abod analysis
}
