% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{comparison}
\alias{comparison}
\title{Do the comparison}
\usage{
comparison(fit, contrast)
}
\arguments{
\item{fit}{the glmfit form edgeR}

\item{contrast}{the contrast data.table}
}
\value{
list with 3 elements:
\itemize{
\item data, a data.table with:
\itemize{
\item rn: the name of the genes
\item logFC: log fold change
\item pval_adj: pvalue adjusted
\item comp_name: the name of the comparison
}
\item plot, a list with the names of the comparison, inside each of them there is another list with:
\itemize{
\item the Smear plot (mean-difference plot)
\item volcano plot
}
}
}
\description{
Make all the comparison that is inside the contrast arguments
}
