% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{dbscan_analysis}
\alias{dbscan_analysis}
\title{dbscan analysis}
\usage{
dbscan_analysis(data, scan = NULL, epsilon = 0, minpts = 0)
}
\arguments{
\item{data}{the data we want}

\item{scan}{the dbscan if it exist}

\item{epsilon}{for the dbscan}

\item{minpts}{for the dbscan}
}
\description{
dbscan analysis
}
