% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{DE}
\alias{DE}
\title{Differential expression}
\usage{
DE(y, conditions)
}
\arguments{
\item{y}{DGElist normalized}

\item{conditions}{conditions of the analysis}
}
\value{
the glm fit tagwise dispersion
}
\description{
Differential expression
}
