% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{filter}
\alias{filter}
\title{Filter the counting table}
\usage{
filter(counts)
}
\arguments{
\item{counts}{The counting table}
}
\value{
the matrix that had been filtered
}
\description{
Remove all the lines with at least the count per million is superior than 1 on at least 3 columns
}
