% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{isofor_analysis}
\alias{isofor_analysis}
\title{isolation forest}
\usage{
isofor_analysis(data, isofor = NULL, nTrees = 100, phi = 8)
}
\arguments{
\item{data}{for the analize}

\item{isofor}{the object if it exist to not}

\item{nTrees}{see \code{\link[isofor:iForest]{isofor::iForest()}}}

\item{phi}{see \code{\link[isofor:iForest]{isofor::iForest()}}}
}
\description{
isolation forest
}
