% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{kNNdistplot}
\alias{kNNdistplot}
\title{kNNdistplot}
\usage{
kNNdistplot(data, k = 4, eps = 0.5, meanDist = F)
}
\description{
Take a object for view the distance beetween the points with various points
}
