% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_analysis.R
\name{normalization}
\alias{normalization}
\title{Normalize the dataset}
\usage{
normalization(counts, groups)
}
\arguments{
\item{counts}{matrix of the counting table}

\item{groups}{the groups (flowcells) for each of the column of the data}
}
\value{
a list with the normalized dataset in DGElist and the plot associated
}
\description{
Normalize the dataset
}
