% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manipulation.R
\name{plot_list_save}
\alias{plot_list_save}
\title{put list of ggplot in a directory by names}
\usage{
plot_list_save(plot_list, path)
}
\arguments{
\item{plot_list}{the plot list}

\item{path}{the path you want to save the plot}
}
\value{
nothing
}
\description{
put list of ggplot in a directory by names
}
