% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manipulation.R
\name{read_parameter_file}
\alias{read_parameter_file}
\title{Read the parameter file}
\usage{
read_parameter_file(name_file)
}
\arguments{
\item{name_file}{path to the file we went to read}
}
\description{
Read the parameter file
}
