% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_manipulation.R
\name{write_parameter_file}
\alias{write_parameter_file}
\title{Write the parameter file}
\usage{
write_parameter_file(param, name_file)
}
\arguments{
\item{param}{param we want to write}

\item{name_file}{the name of the file}
}
\description{
Write the parameter file
}
