% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{pca_analysis}
\alias{pca_analysis}
\title{PCA}
\usage{
pca_analysis(pca, axis1 = 1, axis2 = 2, radius = 0)
}
\arguments{
\item{pca}{the pca if it exist to not recalculate them}

\item{axis1}{the number of the axis 1}

\item{axis2}{the number of the axis 2}

\item{radius}{the radius of the circle}

\item{data}{the data to to the pca analysis}
}
\description{
Do the analysis + the graph for the pca
}
\examples{
""
}
