% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_data.R
\name{tsne_analysis}
\alias{tsne_analysis}
\title{tsne analysis}
\usage{
tsne_analysis(tsne = NULL, scan = NULL)
}
\arguments{
\item{tsne}{if exist don't reculate it}

\item{scan}{if exist the dbscan of the tsne}

\item{data}{the data set}

\item{epsilon}{for the dbscan}

\item{minpts}{for the minpts}
}
\description{
tsne analysis
}
\examples{
""
}
